requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
openpyxl
pymongo[srv]>=3.11
scikit-learn
dill
pyaml
mlflow
dagshub
fastapi
uvicorn


##-e.